requirements.txt
python-telegram-bot
gspread
python-dotenv
requests
